Change long to size_t

+++ interface.pyx
@@ -12,7 +12,7 @@ cdef extern from "src/haar.hh":
         Haar(vector[unsigned char]& img, size_t width, size_t height) except +
         void genIntegral()
         void scanImage()
-        vector[vector[long]]* getValues()
+        vector[vector[size_t]]* getValues()
 
 def processImage(path):
     cdef Haar haar
